requirements.txt
alembic==1.8.1
aniso8601==7.0.0
arrow==1.2.2
async-timeout==4.0.2
attrs==22.1.0
bidict==0.22.0
blinker==1.5
cachelib==0.9.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==7.1.2
cryptography==37.0.4
Deprecated==1.2.13
Flask==1.1.4
Flask-Cors==3.0.10
Flask-GraphQL==2.0.1
Flask-Mail==0.9.1
flask-marshmallow==0.14.0
Flask-Migrate==3.1.0
Flask-OAuthlib==0.9.6
flask-restx==0.5.1
Flask-SocketIO==5.2.0
Flask-SQLAlchemy==2.5.1
graphene==2.1.8
graphene-sqlalchemy==2.3.0
graphql-core==2.3.2
graphql-relay==2.0.1
graphql-server-core==1.2.0
greenlet==1.1.2
idna==3.3
importlib-metadata==4.12.0
importlib-resources==5.9.0
itsdangerous==1.1.0
Jinja2==2.11.3
jsonschema==4.9.1
jwt==1.3.1
Mako==1.2.1
MarkupSafe==2.0.1
marshmallow==3.14.1
marshmallow-sqlalchemy==0.28.1
oauthlib==2.1.0
packaging==21.3
pkgutil_resolve_name==1.3.10
promise==2.3
psycopg2-binary==2.8.6
pycparser==2.21
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==0.20.0
python-engineio==4.3.4
python-socketio==5.7.1
pytz==2022.2.1
redis==4.3.4
requests==2.28.1
requests-oauthlib==1.1.0
rq==1.11.0
rq-dashboard==0.6.1
Rx==1.6.1
singledispatch==3.7.0
six==1.16.0
SQLAlchemy==1.4.40
stripe==4.0.2
urllib3==1.26.11
Werkzeug==1.0.1
wrapt==1.14.1
zipp==3.8.1
